src/mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server providing function calling capabilities via HTTP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.100.0",
]
